+++ types/three/src/Three.WebGPU.d.ts
@@ -8,8 +8,8 @@ export { default as PostProcessing } from "./renderers/common/PostProcessing.js"
 export { default as QuadMesh } from "./renderers/common/QuadMesh.js";
 export type { default as Renderer } from "./renderers/common/Renderer.js";
 export { default as WebGPURenderer } from "./renderers/webgpu/WebGPURenderer.js";
-import * as PostProcessingUtils from "./renderers/common/PostProcessingUtils.js";
-export { PostProcessingUtils };
+import * as RendererUtils from "./renderers/common/RendererUtils.js";
+export { RendererUtils };
 export { default as IESSpotLight } from "./lights/webgpu/IESSpotLight.js";
 export { default as NodeLoader } from "./loaders/nodes/NodeLoader.js";
 export { default as NodeMaterialLoader } from "./loaders/nodes/NodeMaterialLoader.js";
